refactor: split up settings view into multiple subviews
